// frontend/src/backend-types/api.ts
export interface Movie {
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    'id'?: string;
    /**
     * 
     * @type {number}
     * @memberof Movie
     */
    'average_rating'?: number;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    'created_at'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    'updated_at'?: string;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    'rated'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    'released'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Movie
     */
    'runtime'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    'plot'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    'poster'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    'imdbrating'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Movie
     */
    'imdbvotes'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Movie
     */
    'imdbid'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Movie
     */
    'boxoffice'?: number | null;
    /**
     * 
     * @type {Set<string>}
     * @memberof Movie
     */
    'genres'?: Set<string>;
    /**
     * 
     * @type {Set<string>}
     * @memberof Movie
     */
    'awards'?: Set<string>;
    /**
     * 
     * @type {Set<string>}
     * @memberof Movie
     */
    'countries'?: Set<string>;
    /**
     * 
     * @type {Set<string>}
     * @memberof Movie
     */
    'languages'?: Set<string>;
    /**
     * 
     * @type {Set<string>}
     * @memberof Movie
     */
    'actors'?: Set<string>;
    /**
     * 
     * @type {Set<string>}
     * @memberof Movie
     */
    'directors'?: Set<string>;
    /**
     * 
     * @type {Set<string>}
     * @memberof Movie
     */
    'writers'?: Set<string>;
}